Moved tests that connect to Graphviz to integration tests
